Changed: adjustements in test specs
